fix: bug#1261 fixed the crash when some projects hasn't group asigned
The main problem comes from the projects created before the plugin is
installed. They have no group (a project custom field) asigned so it
fails trying to show the metrics by groups.
But actually, we can't do the metrics for those projects because need
all the obligatory custom fields filled.